rename tag queryhandler files, prep for ocp+aws tag query handler (#865)
