+++ koku/api/tags/test/ocp/test_ocp_tag_query_handler.py
@@ -20,7 +20,7 @@ from tenant_schemas.utils import tenant_context
 from api.functions import JSONBObjectKeys
 from api.iam.test.iam_test_case import IamTestCase
 from api.report.test.ocp.helpers import OCPReportDataGenerator
-from api.tags.ocp.ocp_tag_query_handler import OCPTagQueryHandler
+from api.tags.ocp.queries import OCPTagQueryHandler
 from api.utils import DateHelper
 from reporting.models import OCPStorageLineItemDailySummary, OCPUsageLineItemDailySummary
